refactor(marketing): split the vocabulary, derivation and playbook contract out (EP-5CC9C184)
Pure extraction to stay inside the module-size ceiling. No behaviour changes;
330 marketing tests pass unchanged.

lib/marketing.ts was doing three jobs at 1393 LOC: deriving a first strategy
from BusinessContext, normalizing persisted JSON back into shape, and formatting
it for display. The seed-segment change pushed it past its baseline, and the
right answer to a ratchet is to split the module, never to raise the ceiling.

  - lib/marketing/vocabulary.ts   the closed enums plus cleanText/dedupeStrings
  - lib/marketing/strategy-derivation.ts   the infer*/build* family
  - lib/tak/marketing-playbook-types.ts    the playbook contract, split from the
    category table it constrains

Every extracted name is re-exported from its old home, so no consumer changes.
The vocabulary split is what removes the cycle: strategy-derivation needs
MARKETING_CHANNELS, and taking it from lib/marketing.ts — the module that
imports strategy-derivation — would have made a runtime import cycle that
happens to work only because the constant is read inside a function body.

Process-Spine-Decision: extraction only; every name keeps its published import
path through a re-export, so no contract or doc surface changed.
Design-Grounding-Decision: no-contract-change (module split to satisfy the
size ratchet; every extracted name re-exported from its previous home)
